requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aniso8601==9.0.1
    # via graphene
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
authlib==1.3.1
    # via drf-oidc-auth
azure-core==1.30.2
    # via
    #   azure-storage-blob
    #   django-storages
azure-storage-blob==12.20.0
    # via django-storages
cachetools==5.4.0
    # via
    #   django-helusers
    #   google-auth
certifi==2024.7.4
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.8
    # via
    #   authlib
    #   azure-storage-blob
    #   drf-oidc-auth
deprecation==2.1.0
    # via django-helusers
django==4.2.14
    # via
    #   -r requirements.in
    #   django-anymail
    #   django-cors-headers
    #   django-filter
    #   django-guardian
    #   django-helusers
    #   django-ilmoitin
    #   django-mailer
    #   django-parler
    #   django-storages
    #   djangorestframework
    #   drf-oidc-auth
    #   drf-spectacular
    #   graphene-django
    #   helsinki-profile-gdpr-api
django-anymail==11.0.1
    # via django-ilmoitin
django-cleanup==8.1.0
    # via -r requirements.in
django-cors-headers==4.4.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
django-guardian==2.4.0
    # via -r requirements.in
django-helusers==0.13.0
    # via
    #   -r requirements.in
    #   helsinki-profile-gdpr-api
django-ilmoitin==0.7.0
    # via -r requirements.in
django-mailer==2.3.2
    # via django-ilmoitin
django-parler==2.3
    # via
    #   -r requirements.in
    #   django-ilmoitin
django-storages[azure,google]==1.14.4
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-oidc-auth
    #   drf-spectacular
    #   helsinki-profile-gdpr-api
drf-oidc-auth==3.0.0
    # via helsinki-profile-gdpr-api
drf-spectacular==0.27.2
    # via -r requirements.in
ecdsa==0.19.0
    # via python-jose
factory-boy==3.3.0
    # via -r requirements.in
faker==26.0.0
    # via factory-boy
google-api-core==2.19.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.32.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.17.0
    # via django-storages
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.1
    # via google-cloud-storage
googleapis-common-protos==1.63.2
    # via google-api-core
graphene==3.3
    # via graphene-django
graphene-django==3.2.2
    # via -r requirements.in
graphene-file-upload==1.3.0
    # via -r requirements.in
graphql-core==3.2.3
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.2.0
    # via
    #   graphene
    #   graphene-django
hashids==1.3.1
    # via -r requirements.in
helsinki-profile-gdpr-api==0.2.0
    # via -r requirements.in
idna==3.7
    # via requests
importlib-metadata==8.0.0
    # via markdown
inflection==0.5.1
    # via drf-spectacular
isodate==0.6.1
    # via azure-storage-blob
jinja2==3.1.4
    # via django-ilmoitin
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
lockfile==0.12.2
    # via django-mailer
markdown==3.6
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
packaging==24.1
    # via deprecation
pillow==10.4.0
    # via -r requirements.in
promise==2.3
    # via graphene-django
proto-plus==1.24.0
    # via google-api-core
protobuf==5.27.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2==2.9.9
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycountry==24.6.1
    # via -r requirements.in
pycparser==2.22
    # via cffi
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.9.0.post0
    # via faker
python-jose==3.3.0
    # via django-helusers
pytz==2024.1
    # via -r requirements.in
pyyaml==6.0.1
    # via drf-spectacular
qrcode==7.4.2
    # via -r requirements.in
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   azure-core
    #   django-anymail
    #   django-helusers
    #   drf-oidc-auth
    #   google-api-core
    #   google-cloud-storage
rpds-py==0.19.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via
    #   google-auth
    #   python-jose
sentry-sdk==2.10.0
    # via -r requirements.in
six==1.16.0
    # via
    #   azure-core
    #   ecdsa
    #   graphene-file-upload
    #   isodate
    #   promise
    #   python-dateutil
sqlparse==0.5.1
    # via django
text-unidecode==1.3
    # via graphene-django
typing-extensions==4.12.2
    # via
    #   asgiref
    #   azure-core
    #   azure-storage-blob
    #   drf-spectacular
    #   qrcode
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.2
    # via
    #   django-anymail
    #   requests
    #   sentry-sdk
zipp==3.19.2
    # via importlib-metadata
